move rides to routes
