+++ src/routes/index.js
@@ -1,7 +1,8 @@
 const app = module.exports = require('express')();
 
-module.exports = (db, logger) => {
+module.exports = (db) => {
   app.use('/health', require('./health'));
+  app.use('/rides', require('./rides')(db));
 
   // the catch all route
   app.all('*', (req, res) => {
